ContentfulRichText: move Node into nodes directory
